% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leading_zeroes_and_letter.R
\name{add_leading_zeroes_and_letter}
\alias{add_leading_zeroes_and_letter}
\title{Add a column 'number' that counts entries in order of date -- adds leading zero, and optional a letter}
\usage{
add_leading_zeroes_and_letter(cv_entries_tibble, letter_to_pad = "")
}
\arguments{
\item{cv_entries_tibble}{A tibble that has a column called 'date' which contains dates!}

\item{letter_to_pad}{A letter that should go in front of the counted entries}
}
\value{
A tibble with an added 'number' column
}
\description{
Add a column 'number' that counts entries in order of date -- adds leading zero, and optional a letter
}
\examples{
publications |>
  filter(type == "journal article") |>
  add_leading_zeroes_and_letter("J")

}
