% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_year_when_repeated.R
\name{blank_year_when_repeated}
\alias{blank_year_when_repeated}
\title{Turn the year into an empty string when it's the same year as the previous entry}
\usage{
blank_year_when_repeated(cv_entry_tibble)
}
\arguments{
\item{cv_entry_tibble}{A tibble with a column called 'year' that contains years...}
}
\value{
A tibble with an added 'number' column
}
\description{
Turn the year into an empty string when it's the same year as the previous entry
}
\examples{
publications |>
  filter(type == "journal article") |>
  blank_year_when_repeated()

}
