% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_page_break_after_row.R
\name{manual_page_break_after_row}
\alias{manual_page_break_after_row}
\title{Adds a page break after a row that contains a specified text}
\usage{
manual_page_break_after_row(
  some_tibble,
  a_column_name,
  text_to_detect,
  use_glue = TRUE
)
}
\arguments{
\item{some_tibble}{A tibble}

\item{a_column_name}{The name of the column where you'd like to detect text}

\item{text_to_detect}{The text you'd like to detect -- a page break is added by the end of the rows with this text}

\item{use_glue}{Whether to have the output be of type glue/character (default), or character}
}
\value{
A tibble with a page break added by the end of specified
}
\description{
Adds a page break after a row that contains a specified text
}
\examples{
# add page break after the GDPR paper
publications |>
 filter(type == "journal article") |>
 mutate(citation = glue::glue("{title}, {year}")) |>
 select(year, citation) |>
 manual_page_break_after_row("citation", "Before and after GDPR")

}
