% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\docType{data}
\name{publications}
\alias{publications}
\title{Publication entries from Ulrik Lyngs' CV}
\format{
A tibble with 26 rows and 30 variables:
\describe{
  \item{type}{chr The type of publication: journal article, workshop paper, conference paper, etc.}
  \item{authors}{chr The authors}
  \item{date}{date The publication date}
  \item{title}{chr The publication title}
  \item{abstract}{chr The publication abstract}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1ta71CAGkcLqm-W1UdVRA_JJSddWV2TsrRZsCnQlmOis/edit?usp=sharing}
}
\usage{
publications
}
\description{
Publication entries from Ulrik Lyngs' CV
}
\keyword{datasets}
