% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_first_name_with_initial_and_dot.R
\name{replace_first_name_with_initial_and_dot}
\alias{replace_first_name_with_initial_and_dot}
\title{Take an author name and replace the first name with an initial and dot}
\usage{
replace_first_name_with_initial_and_dot(author_name)
}
\arguments{
\item{authors}{A string with an author name (e.g., "Ulrik Lyngs")}
}
\value{
When given "Ulrik Lyngs" it returns "U. Lyngs"
}
\description{
Take an author name and replace the first name with an initial and dot
}
\examples{
replace_first_name_with_initial_and_dot("Ulrik Lyngs")

}
