% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/replace_first_names_in_author_list_with_initial_and_dot.R
\name{replace_first_names_in_author_list_with_initial_and_dot}
\alias{replace_first_names_in_author_list_with_initial_and_dot}
\title{Take a comma-separated string of authors and replace the first names with an initial and a dot}
\usage{
replace_first_names_in_author_list_with_initial_and_dot(authors)
}
\arguments{
\item{authors}{A comma-separated string of authors (e.g., "Ulrik Lyngs, Kai Lukoff")}
}
\value{
When given "Ulrik Lyngs, Kai Lukoff" it returns "U. Lyngs, K. Lukoff"
}
\description{
Take a comma-separated string of authors and replace the first names with an initial and a dot
}
\examples{
replace_first_names_in_author_list_with_initial_and_dot("Ulrik Lyngs, Kai Lukoff")

}
