% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gsheet_as_csv.R
\name{save_gsheet_as_csv}
\alias{save_gsheet_as_csv}
\title{Read in a public google sheet and save it as CSV}
\usage{
save_gsheet_as_csv(
  sheet_name,
  gsheet_url =
    "https://docs.google.com/spreadsheets/d/1ta71CAGkcLqm-W1UdVRA_JJSddWV2TsrRZsCnQlmOis/edit?usp=sharing",
  store_path = paste0(sheet_name, ".csv")
)
}
\arguments{
\item{sheet_name}{The name of a single sheet within the google sheet}

\item{gsheet_url}{A URL to a google sheet. By default uses Ulrik Lyngs' CV google sheet. Create a link to your own google sheet by clicking the 'Share' button on a google sheet, then create a general access link.}

\item{store_path}{Where to store the downloaded sheet. By default stores it as a CSV file with the same name as the sheet.}
}
\value{
Nothing
}
\description{
Read in a public google sheet and save it as CSV
}
\examples{
my_gsheet <- "https://docs.google.com/spreadsheets/d/1ta71CAGkcLqm-W1UdVRA_JJSddWV2TsrRZsCnQlmOis/edit?usp=sharing"
save_gsheet_as_csv("cv_entries", my_gsheet)

}
