% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\docType{data}
\name{publications}
\alias{publications}
\title{Publication entries from Ulrik Lyngs' CV}
\format{
A tibble with 26 rows and 10 variables:
\describe{
  \item{type}{chr The type of publication: journal article, workshop paper, conference paper, etc.}
  \item{authors}{chr The authors}
  \item{year}{dbl The year of the publication}
  \item{date}{date The exact publication date}
  \item{title}{chr The publication title}
  \item{venue}{chr The venue of the publication}
  \item{venue_abbrev}{chr An abbreviation of the venue}
  \item{award}{chr An award received for the publication}
  \item{url}{chr Link to the publication}
  \item{materials}{chr Link to the materials used in the publication}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1ta71CAGkcLqm-W1UdVRA_JJSddWV2TsrRZsCnQlmOis/edit?usp=sharing}
}
\usage{
publications
}
\description{
Publication entries from Ulrik Lyngs' CV
}
\keyword{datasets}
